Hipaa-training
The Health Insurance Portability and Accountability Act, usually known as HIPAA, outlines patient privacy laws within the US healthcare system and requires that anyone accessing or managing healthcare information needs HIPAA training. As technology evolves, the law and its applicability are ever-changing. Even those employees that already have experience dealing with HIPAA may find that they are not up-to-date with current best practices.

That is where training comes in. A well-designed training course can save an organization time and money, as employees will make fewer mistakes that cost the organization. Human error remains one of the foremost threats to patient privacy – a threat that can easily be remedied through training.

HIPAA is complex and nuanced, and thus it is unreasonable to believe that any one regular employee will remember its every detail. However, no employee that deals with patient data – be they a doctor, accountant, actuary or nurse – should begin work without being acquainted with HIPAA policies and practices. As most of these employees will not have law degrees, training courses ensure that employees receive appropriate information without trying to interpret the document themselves. This is especially important given the broad nature of HIPAA; as it was written to apply to many different types of organizations and cover many different scenarios. There are also many requirements that are nuanced, which can lead to confusion. Interpreting the legislation should be left to those with proper training.

Despite the obvious need for training, the actual training guidelines stipulated by HIPAA are limited. Under the HIPAA Privacy Rule, training should be provided “as necessary and appropriate for members of the workforce to carry out their functions”, whilst the Security Rule requires organizations to  “implement a security awareness and training program for all members of the workforce.” Aside from these two requirements, and the need to provide training when there is a change in technology, policies, or updates to the HIPAA Rules, no other information is provided about the content and frequency of training.